pyproject.toml
[project]
name = "dojoflow"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.4.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.3.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
